fix!: digital contract guards alerts when fields are not complete (#4)
now the digital contract generation has some alerts and some guards like
exceptions when the data is not sufficient. We need to address other
problems that might arise from this, but at the moment is working fine.

BREAKING CHANGE: you can have some problems with the Contacto and User
classes, as now some methods throw exceptions if, for example, the
contract is not found, so it will send a ModelNotFoundException.

fixes #4